UD.SACO Spring Rest API @DeleteMapping, @PathVariable l.87

+++ src/main/java/com/vitaliichekalenko/spring/rest/controller/MyRESTController.java
@@ -56,6 +56,11 @@ public class MyRESTController {
     @DeleteMapping("/employees/{id}")
     public String deleteEmployee(@PathVariable int id){
 
+        Employee employee = employeeService.getEmployee(id);
+        if(employee==null){
+            throw new NoSuchEmployeeException("There is no employee where id = " + id + " in database");
+        }
+
         employeeService.deleteEmployee(id);
         return "Employee with id = " + id + " was deleted";
